+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/aad/webapi/AADJwtBearerTokenAuthenticationConverter.java
@@ -4,6 +4,8 @@ package com.azure.spring.aad.webapi;
 
 import com.azure.spring.aad.AADOAuth2AuthenticatedPrincipal;
 import com.azure.spring.aad.AbstractJwtBearerTokenAuthenticationConverter;
+import com.azure.spring.aad.implementation.constants.AADTokenClaim;
+import com.azure.spring.aad.implementation.constants.AuthorityPrefix;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.OAuth2AuthenticatedPrincipal;
@@ -14,8 +16,6 @@ import java.util.Collection;
 import java.util.Map;
 import java.util.Optional;
 
-import static com.azure.spring.aad.AADJwtGrantedAuthoritiesConverter.DEFAULT_AUTHORITY_PREFIX;
-import static com.azure.spring.aad.AADJwtGrantedAuthoritiesConverter.DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP;
 
 /**
  * A {@link Converter} that takes a {@link Jwt} and converts it into a {@link BearerTokenAuthentication}.
@@ -23,10 +23,10 @@ import static com.azure.spring.aad.AADJwtGrantedAuthoritiesConverter.DEFAULT_CLA
 public class AADJwtBearerTokenAuthenticationConverter extends AbstractJwtBearerTokenAuthenticationConverter {
 
     /**
-     * Construct AADJwtBearerTokenAuthenticationConverter by DEFAULT_PRINCIPAL_CLAIM_NAME and DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP.
+     * Construct AADJwtBearerTokenAuthenticationConverter by AADTokenClaim.SUB and DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP.
      */
     public AADJwtBearerTokenAuthenticationConverter() {
-        this(DEFAULT_PRINCIPAL_CLAIM_NAME, DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP);
+        this(AADTokenClaim.SUB, AADResourceServerProperties.DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP);
     }
 
     /**
@@ -34,7 +34,7 @@ public class AADJwtBearerTokenAuthenticationConverter extends AbstractJwtBearerT
      * @param authoritiesClaimName authority claim name
      */
     public AADJwtBearerTokenAuthenticationConverter(String authoritiesClaimName) {
-        this(authoritiesClaimName, DEFAULT_AUTHORITY_PREFIX);
+        this(authoritiesClaimName, AuthorityPrefix.SCOPE);
     }
 
     /**